valueOfConverters already in codebase
git-svn-id: http://svn.codehaus.org/picocontainer/java/2.x/trunk@5198 ac66bb80-72f5-0310-8d68-9f556cfffb23

+++ pico/container/src/java/org/picocontainer/parameters/BasicComponentParameter.java
@@ -104,7 +104,7 @@ public class BasicComponentParameter implements Parameter, Serializable {
         stringConverters.put(Float.class, new ValueOfConverter(Float.class));
         stringConverters.put(Character.class, new ValueOfConverter(Character.class));
         stringConverters.put(Byte.class, new ValueOfConverter(Byte.class));
-        stringConverters.put(Byte.class, new ValueOfConverter(Short.class));
+        stringConverters.put(Short.class, new ValueOfConverter(Short.class));
         stringConverters.put(File.class, new NewInstanceConverter(File.class));
 
     }
